+++ src/components/PokeSearchForm/PokemonList/PokemonList.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { Outlet } from 'react-router-dom';
-import { PokemonData, PokemonResult } from '../../types/types';
-import { stopPropagation } from '../../utils/eventHandler/stopPropaganation';
-import Spinner from '../spinner/Spinner';
-import PokeApi from '../../services/PokeApi';
-import PokemonCard from './PokemonCard';
+import { PokemonData, PokemonResult } from '../../../types/types';
+import { stopPropagation } from '../../../utils/eventHandler/stopPropaganation';
+import Spinner from '../../spinner/Spinner';
+import PokeApi from '../../../services/PokeApi';
+import PokemonCard from '../PokemonCard/PokemonCard';
 
 import './PokemonList.css';
